--- a/web/api/v1/api.go
+++ b/web/api/v1/api.go
@@ -24,6 +24,8 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/mailru/easyjson"
+	"github.com/mailru/easyjson/jwriter"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/common/route"
@@ -78,6 +80,7 @@ type alertmanagerRetriever interface {
 	Alertmanagers() []*url.URL
 }
 
+//easyjson:json
 type response struct {
 	Status    status      `json:"status"`
 	Data      interface{} `json:"data,omitempty"`
@@ -152,6 +155,28 @@ func (api *API) Register(r *route.Router) {
 type queryData struct {
 	ResultType model.ValueType `json:"resultType"`
 	Result     model.Value     `json:"result"`
+}
+
+// MarshalJSON supports json.Marshaler interface
+func (v queryData) MarshalJSON() ([]byte, error) {
+	w := jwriter.Writer{}
+	v.MarshalEasyJSON(&w)
+	return w.Buffer.BuildBytes(), w.Error
+}
+
+// MarshalEasyJSON supports easyjson.Marshaler interface
+func (v queryData) MarshalEasyJSON(out *jwriter.Writer) {
+	out.RawByte('{')
+	out.RawString("\"resultType\":\"")
+	out.RawString(v.ResultType.String())
+	out.RawString(`",`)
+	out.RawString("\"result\":")
+	if m, ok := v.Result.(easyjson.Marshaler); ok {
+		m.MarshalEasyJSON(out)
+	} else {
+		out.Raw(json.Marshal(v.Result))
+	}
+	out.RawByte('}')
 }
 
 func (api *API) options(r *http.Request) (interface{}, *apiError) {
@@ -437,17 +462,17 @@ func (api *API) alertmanagers(r *http.Request) (interface{}, *apiError) {
 }
 
 func respond(w http.ResponseWriter, data interface{}) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-
-	b, err := json.Marshal(&response{
+	r := response{
 		Status: statusSuccess,
 		Data:   data,
-	})
-	if err != nil {
-		return
 	}
-	w.Write(b)
+
+	w.Header().Set("Content-Type", "application/json")
+
+	started, _, err := easyjson.MarshalToHTTPResponseWriter(r, w)
+	if err != nil && !started {
+		respondError(w, &apiError{errorInternal, err}, nil)
+	}
 }
 
 func respondError(w http.ResponseWriter, apiErr *apiError, data interface{}) {