This is a fix for #3601, this wouldn't be the *actual* PR (as we'll need to commit the changes in https://github.com/prometheus/common/pull/112 first) but this gives an idea of the level of change required (very small).

I've run this locally with some test data, and here are the comparisons:

|        | pre-patch | with-patch |
|--------|-----------|------------|
| Matrix | 214ms     | 37ms       |
| Vector | 108ms     | 20ms       |
| Scalar | 20ms      | 11ms       |
| String | 18ms      | 13ms       |

With this we can see the large performance gains even with small returns. Not only do we get these gains (~5-6 for "real" data) we also significantly reduce the memory required to marshal out the response (which means this scales *significantly* better into large response sizes.

Also, as a bonus, since this iteratively serializes it will also *stop* mid-serialization if the client disconnects!