@jacksontj 2017-12-28
To include a complete end-to-end benchmark I've done a simple promql query (`metricname[range]`) over various time durations pre and post easyjson change, here are the results:

| Time (minutes) | Old    | New   | Improvement  |
|----------------|--------|-------|--------------|
| 0              | 0.121  | 0.086 | -28.92561983 |
| 1              | 0.171  | 0.076 | -55.55555556 |
| 10             | 1.226  | 0.119 | -90.29363785 |
| 30             | 2.126  | 0.58  | -72.7187206  |
| 60             | 5.173  | 0.441 | -91.47496617 |
| 120            | 9.086  | 1.107 | -87.81642087 |
| 180            | 13.719 | 1.795 | -86.91595597 |
| 240            | 17.668 | 1.956 | -88.92913742 |