@brian-brazil 2017-12-28
This PR is against Prometheus 1.7, which is not the current development version. Any changes of this nature should be proposed against master.