@jacksontj 2017-12-28
@brian-brazil I can do that, I opened the PR here as the issue (#3601 ) was reported on 1.7. The changes here can (and should) be applied to any releases that are still supported (I'm not sure which ones those are, so I'll need some guidance on what are the currently supported releases).