@brian-brazil 2017-12-28
A change like this would be aimed for 2.1 currently, which is to say master. New features are not being added to 1.x.