@jacksontj 2017-12-28
I can re-open the PR against master, but https://github.com/prometheus/common/pull/112 needs to be merged first. Before moving the PR I'd like to see that merged and some feedback on this change (since it will be basicaly the same change against master).

Is there also no mechanism to backport to previous releases? From looking at the releases page it seems that N-1 is supported (minor releases etc.)?