@jacksontj 2017-12-28
What classifies bug fix? From the end to end data this is ~10x perf
improvement.

Also, your latest response still has no feedback for a change like this.

On Dec 27, 2017 4:48 PM, "Brian Brazil" <notifications@github.com> wrote:

> All the relevant code changed with 2.0, you'd need to redo it all.
>
> Only bugfixes would be considered for backport.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/prometheus/pull/3608#issuecomment-354205590>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABi3nXHiLCCYI_2QTp4GzjKfuVV9Tt1Zks5tEuVhgaJpZM4RKaNd>
> .
>