@brian-brazil 2017-12-28
A performance improvement is not considered a bugfix.