@jacksontj 2017-12-28
Ok, how about feedback on the change?

On Dec 27, 2017 4:57 PM, "Brian Brazil" <notifications@github.com> wrote:

> A performance improvement is not considered a bugfix.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/prometheus/pull/3608#issuecomment-354206788>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABi3nR1QakyNWtnQ3j9B1mWlfCAoIqyuks5tEudsgaJpZM4RKaNd>
> .
>