@brian-brazil 2017-12-28
It's a highly invasive change that'd require significant ongoing error-prone maintenance, and it also violates layering. Such downsides require major major benefits to justify, and the benchmark numbers produced thus far are at odds with previous profiling in this area.