@jacksontj 2017-12-28
> It's a highly invasive change 

I'm not sure if we're looking at the same diff? The reason this particular PR is so large is due to changes in the vendor directory. The actual change to the code is ~30-40 lines of change, which I'd hardly call "highly invasive".

> that'd require significant ongoing error-prone maintenance

I'm also not sure what maintenance you are expecting to perform? There is one codegen'd file -- which requires no human to write the code and the only other one is a small marshal method on the queryData struct -- which isn't exactly a large struct.

> and it also violates layering.

What are you implying with this? What layering is being broken? In golang marshal methods have to be in the package with the struct definition-- so if you don't want the marshal/unmarshal methods in this package then we need to move it to another package. Past that I have no idea what other layering you could be referring to, so please elaborate.

> Such downsides require major major benefits to justify

I agree, which is why I've spent a significant amount of time collecting a LOT of data (and creating a variety of benchmarks to demonstrate this)

> and the benchmark numbers produced thus far are at odds with previous profiling in this area.

Thats usually how performance improvements work? For all the benchmarks we're talking about the code is all open -- so it is comparable. The benchmarks you are referring to are all done by the jsoniter guys on a generic dataset. The benchmarks I've put in place are (1) with the prometheus structs and (2) a full end-to-end test (which is exactly what you asked for). At this point I'm at a loss for what additional information to provide. I have overwhelming data supporting my position, and all you've been able to counter with is "suspicion" -- with no recourse.

What data or proof do you require to get such a change in? I'm growing weary of gathering data only to be met with 'suspicion' and uncertainty. This is a simple performance optimization (a HUGE one, but a simple one nonetheless). If not performance data, what data do you need?